+++ src/types/property-types/computed-property.ts
@@ -17,7 +17,7 @@ export class ComputedProperty extends Property {
 
     validate(snapshot: any, context: IContext): IValidationResult {
         if ( this.name in snapshot ) {
-            return [{ context: context.concat([ { path: this.name } ]), snapshot: snapshot[this.name], message: "Computed properties should not be provided in the snapshot" }]
+            return [{ context: context.concat([ { path: this.name } ]), value: snapshot[this.name], message: "Computed properties should not be provided in the snapshot" }]
         }
 
         return []