+++ src/types/property-types/view-property.ts
@@ -18,7 +18,7 @@ export class ViewProperty extends Property {
 
     validate(snapshot: any, context: IContext): IValidationResult {
         if( this.name in snapshot ) {
-            return [{ context: context.concat([ { path: this.name } ]), snapshot: snapshot[this.name], message: "View properties should not be provided in the snapshot" }]
+            return [{ context: context.concat([ { path: this.name } ]), value: snapshot[this.name], message: "View properties should not be provided in the snapshot" }]
         }
 
         return []